+++ src/components/Card/Card.test.tsx
@@ -13,23 +13,8 @@ import { Provider } from 'react-redux';
 import { store } from '../../store/store';
 import { server } from '../../test/server';
 import { rest } from 'msw';
-// import { AppContext } from '../../context';
 
 describe('Card component', async () => {
-  // const mockedUseFetching = vi.hoisted(() => vi.fn());
-
-  // vi.mock('../../hooks/useFetching', () => ({
-  //   useFetching: mockedUseFetching,
-  // }));
-
-  // beforeEach(() => {
-  //   vi.clearAllMocks();
-  // });
-
-  // afterEach(() => {
-  //   vi.clearAllMocks();
-  // });
-
   beforeAll(() => server.listen());
   afterEach(() => server.resetHandlers());
   afterAll(() => server.close());
@@ -49,21 +34,7 @@ describe('Card component', async () => {
   });
 
   it('Validate that clicking on a card opens a detailed card component', async () => {
-    // mockedUseFetching.mockImplementation(() => ({
-    //   fetchCardById: vi.fn(),
-    //   loading: true,
-    // }));
     render(
-      // <AppContext.Provider
-      //   value={{
-      //     items: [],
-      //     searchValue: '',
-      //     delailsData: detailedMockData,
-      //     setDetailsData: vi.fn,
-      //     setItems: vi.fn,
-      //     setSearchValue: vi.fn,
-      //   }}
-      // >
       <Provider store={store}>
         <MemoryRouter initialEntries={['/']}>
           <Routes>
@@ -72,13 +43,18 @@ describe('Card component', async () => {
           </Routes>
         </MemoryRouter>
       </Provider>
-      // </AppContext.Provider>
     );
 
     await act(async () => {
       const cardContainer = screen.getByTestId('card-container');
       fireEvent.click(cardContainer);
     });
+
+    server.use(
+      rest.get('https://api.jikan.moe/v4/manga/1', (_, res, ctx) => {
+        return res(ctx.status(200), ctx.json(detailedMockData));
+      })
+    );
 
     await waitFor(() => {
       const cardContainer = screen.getByTestId('details');
@@ -87,22 +63,7 @@ describe('Card component', async () => {
   });
 
   it('Check that clicking triggers an additional API call to fetch detailed information', async () => {
-    // const mockFetchCard = vi.fn();
-    // mockedUseFetching.mockImplementationOnce(() => ({
-    //   fetchCardById: mockFetchCard,
-    // }));
-
     render(
-      // <AppContext.Provider
-      //   value={{
-      //     items: [],
-      //     searchValue: '',
-      //     delailsData: detailedMockData,
-      //     setDetailsData: vi.fn,
-      //     setItems: vi.fn,
-      //     setSearchValue: vi.fn,
-      //   }}
-      // >
       <Provider store={store}>
         <MemoryRouter initialEntries={['/']}>
           <Routes>
@@ -111,11 +72,10 @@ describe('Card component', async () => {
           </Routes>
         </MemoryRouter>
       </Provider>
-      // </AppContext.Provider>
     );
 
     server.use(
-      rest.get('https://api.jikan.moe/v4/manga/1', (req, res, ctx) => {
+      rest.get('https://api.jikan.moe/v4/manga/1', (_, res, ctx) => {
         return res(ctx.status(200), ctx.json(detailedMockData));
       })
     );
