+++ src/components/CardList/CardList.test.tsx
@@ -4,8 +4,6 @@ import { CardList } from './CardList';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from '../../store/store';
-// import { detailedMockData } from '../../test/mock_data';
-// import { AppContext } from '../../context';
 
 const items = [
   {
@@ -21,22 +19,11 @@ describe('CardList component', () => {
   it('renders the specified number of cards', async () => {
     await act(async () => {
       render(
-        // <AppContext.Provider
-        //   value={{
-        //     items,
-        //     searchValue: '',
-        //     delailsData: undefined,
-        //     setDetailsData: vi.fn,
-        //     setItems: vi.fn,
-        //     setSearchValue: vi.fn,
-        //   }}
-        // >
         <Provider store={store}>
           <MemoryRouter initialEntries={['/']}>
             <CardList cards={items} />
           </MemoryRouter>
         </Provider>
-        // </AppContext.Provider>
       );
     });
 
@@ -50,20 +37,9 @@ describe('CardList component', () => {
   it('Check that an appropriate message is displayed if no cards are present', async () => {
     await act(async () => {
       render(
-        // <AppContext.Provider
-        //   value={{
-        //     items: [],
-        //     searchValue: '',
-        //     delailsData: undefined,
-        //     setDetailsData: vi.fn,
-        //     setItems: vi.fn,
-        //     setSearchValue: vi.fn,
-        //   }}
-        // >
         <Provider store={store}>
           <CardList cards={[]} />
         </Provider>
-        // </AppContext.Provider>
       );
     });
