+++ src/components/SelectPageSize/SelectPageSize.test.tsx
@@ -3,14 +3,6 @@ import { SelectPageSize } from './SelectPageSize';
 import { Provider } from 'react-redux';
 import { store } from '../../store/store';
 
-// const renderComponent = () => {
-//   return render(
-//     <Provider store={store}>
-//     <SelectPageSize />
-//     </Provider>
-//   );
-// };
-
 describe('SelectPageSize component', () => {
   test('renders correctly', () => {
     const { getByTestId } = render(
@@ -27,13 +19,4 @@ describe('SelectPageSize component', () => {
       expect(option).toBeInTheDocument();
     });
   });
-
-  // test('calls onInputValueChange with the correct value when select input changes', () => {
-  //   const { getByTestId } = renderComponent();
-  //   const selectInput = getByTestId('select-input');
-
-  //   fireEvent.change(selectInput, { target: { value: '15' } });
-
-  //   expect(mockOnInputValueChange).toHaveBeenCalledWith(15);
-  // });
 });
