+++ src/test/server.ts
@@ -3,6 +3,9 @@ import { rest } from 'msw';
 import { detailedMockData, mockData } from './mock_data';
 
 global.fetch = fetch;
+global.Headers = Headers;
+global.Request = Request;
+global.Response = Response;
 
 const server = setupServer(
   rest.get('https://api.jikan.moe/v4/manga', (req, res, ctx) => {
